apd_project/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.14.0
streamlit>=1.28.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
PyYAML>=6.0
